Multiple Step Definitons can be defined now. Test fails because SignUpStepDefs are wrong in this example
