+++ src/test/java/linkedin/testing/pageobjects/SignInPage.java
@@ -5,6 +5,8 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.springframework.stereotype.Component;
 
+import java.util.Map;
+
 @Component
 public class SignInPage extends CommonPageObject{
 
@@ -17,6 +19,11 @@ public class SignInPage extends CommonPageObject{
     @FindBy(id = "password")
     private WebElement passwordInput;
 
+    private final Map<String, WebElement> inputFieldsMap = Map.of(
+            "Email or Phone", emailOrPhone,
+            "Password", passwordInput
+    );
+
     public SignInPage(WebDriverFactory factory) {
         super(factory);
     }
@@ -25,5 +32,9 @@ public class SignInPage extends CommonPageObject{
         waitForElementToBeClickable(signInButton);
         signInButton.click();
         waitForPageReadiness();
+    }
+
+    public WebElement getInputFieldByName(final String name) {
+        return inputFieldsMap.get(name);
     }
 }
