+++ src/test/java/linkedin/testing/stepdefinitions/LinkedinSignInStepDefs.java
@@ -2,7 +2,6 @@ package linkedin.testing.stepdefinitions;
 
 import io.cucumber.java.en.And;
 import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.cucumber.spring.CucumberContextConfiguration;
 import linkedin.testing.config.TestConfig;
@@ -17,20 +16,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
 
 import java.time.Duration;
-import java.util.Collections;
 import java.util.List;
 
 import static linkedin.testing.config.TestConfig.PAGE_OR_ELEMENT_LOAD_WAIT_SECONDS;
 
-@CucumberContextConfiguration
 @ContextConfiguration(classes = TestConfig.class)
-public class LinkedinSignupStepDefs {
+public class LinkedinSignInStepDefs {
 
     @Autowired
     private HomePage homePage;
-
-    @Autowired
-    private JoinNowPage joinNowPage;
 
     @Autowired
     private SignInPage signInPage;
@@ -40,20 +34,9 @@ public class LinkedinSignupStepDefs {
         homePage.navigateToHomePage();
     }
 
-    @And("the Join now header button is clicked")
-    public void theRegistrationHeaderButtonIsClicked() {
-        homePage.clickOnJoinNowHeaderButton();
-    }
-
     @Given("it is scrolled down")
     public void itIsScrolledDown() {
-        joinNowPage.scrollToTheBottomOfThePage();
-    }
-
-    @When("the Agree & Join button is clicked")
-    public void theAgreeAndJoinButtonIsClicked(){
-        System.out.println("aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa");
-        joinNowPage.clickOnAgreeAndJoinButton();
+        signInPage.scrollToTheBottomOfThePage();
     }
 
     @And("the Sign in header button is clicked")
@@ -66,39 +49,11 @@ public class LinkedinSignupStepDefs {
         signInPage.clickOnSignInButton();
     }
 
-    @And("^the '(.*)' error message of the '(?:.*)' (?:field|dropdown|radio buttons|checkbox) should be shown$")
-    public void theErrorMessageShouldBeShown(final String message) {
-        System.out.println("-----------------------------------------");
-        Awaitility.await(String.format("Element was not loaded in %s seconds", PAGE_OR_ELEMENT_LOAD_WAIT_SECONDS))
-                .atMost(Duration.ofSeconds(PAGE_OR_ELEMENT_LOAD_WAIT_SECONDS))
-                .until(() -> joinNowPage.getWebDriverFromFactory().findElements(
-                                By.xpath(String.format("//div[text()=\"%s\" or ./span[text()=\"%s\"]]", message, message))
-                                //By.xpath(String.format("//p[text()=\"%s\" or ./span[text()=\"%s\"]]", message, message))  //linkedin a hibeüzenetet egy <p> elemben írja ki, de így sem jó
-                        ).size(),
-                        Matchers.is(1));
-    }
-
     @And("^the '(.*)' error message should be shown$")
     public void theEmailErrorMessageShouldBeShown(final String message) {
-        System.out.println("bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb");
-        System.out.println("message: " + message);
-        System.out.println(String.format("//div[text()=\"%s\"", message));
-        System.out.println(String.format("//p[@class=\"main-content\"])"));
-        System.out.println("---------------------------");
-/*
-        List<WebElement> elements = joinNowPage.getWebDriverFromFactory().findElements(
-                By.cssSelector("p.artdeco-inline-feedback__message"));
-
-        elements.stream().forEach(e -> System.out.println(e));
-        System.out.println(elements.get(0).getText());
-        System.out.println(elements.get(1).getText());
-        System.out.println("--------------------------");
-        System.out.println(elements.get(0).getTagName());
-        System.out.println(elements.get(1).getTagName());
-*/
         Awaitility.await(String.format("Element was not loaded in %s seconds", PAGE_OR_ELEMENT_LOAD_WAIT_SECONDS))
                 .atMost(Duration.ofSeconds(PAGE_OR_ELEMENT_LOAD_WAIT_SECONDS))
-                .until(() -> joinNowPage.getWebDriverFromFactory().findElements(
+                .until(() -> signInPage.getWebDriverFromFactory().findElements(
                                 By.xpath(String.format("//div[text()=\"%s\" or ./span[text()=\"%s\"]]", message, message)) //eredeti: keresek egy div elemet aminek vagy a szövege egyenlő message-el vagy tartalmaz egy olyan span elemet aminek a szövege egyenlő message-el
                                 //By.xpath(String.format("//p[text()=\"%s\" or ./span[text()=\"%s\"]]", message, message))  //linkedin a hibeüzenetet egy <p> elemben írja ki, de így sem jó
                                 //By.xpath(String.format("//div[text()=\"%s\" or //p[@class=\"artdeco-inline-feedback__message\"])]", message))  //linkedin a hibeüzenetet egy <p> elemben írja ki, de így sem jó
@@ -124,7 +79,6 @@ public class LinkedinSignupStepDefs {
 
     @When("the {string} is filled in with {string}")
     public void theFieldIsFilledWithParameter(final String field, final String content) {
-        joinNowPage.getInputFieldByName(field).sendKeys(content);
+        signInPage.getInputFieldByName(field).sendKeys(content);
     }
-
 }
