+++ src/test/resources/features/linkedin-signin.feature
@@ -1,0 +1,11 @@
+Feature: Linkedin sign in page
+
+  Background:
+    Given the home page is opened
+    And the Sign in header button is clicked
+
+
+  @requiredfield
+  Scenario: Check sign in required fields
+    When the Sign in button is clicked
+    Then the 'Please enter an email address or phone number' error message should be shown
